+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# [Unreleased]
+
+#### New Features
+* Add [`LeelaChessZero (lc0)` engine](https://lczero.org)
+  * Currently comes bundled with a basic neural network
+
 # ChessKitEngine 0.1.3
 Released Saturday, April 15, 2023.